OneDrive/Рабочий стол/skill-swap-main/requirements.txt
Django==5.1.2
django-tailwind==3.8.0
Pillow==10.4.0
djangorestframework==3.15.2
django-cors-headers==4.3.1
gunicorn==21.2.0
whitenoise==6.6.0
psycopg2-binary==2.9.9
dj-database-url==2.1.0
drf-yasg==1.21.7
sentry-sdk==2.15.0
